backend/requirements.txt
fastapi
uvicorn
python-multipart
filetype
slowapi
supabase
python-dotenv
ultralytics
Pillow